Enemy/object rendering capabilities
among general cleanup, and some additional tools in VectorOps, RayCaster received a couple changes to allow Enemy/object rendering capabilities. This has been added in the main rendering loop in protogame as 2 functions to generate visible sprites, and to render them on the correct layer as to stack correctly with and against walls.