important bugfixes for sprites, pygame modularity
I found some flaws in the rendering of sprites, primarily due to an error in VectorOps. These have all been resolved to my knowledge.

Brother Manley wanted pygame to be more encapsulated into it's own environment, so that in theory it would be more modular to swap in and out other renderers. A new class has been made for all pygame draw functions.