Raycast optimization, Pygame Refactoring, HUD Foundation
Massive performance and accuracy uptick for raycast performance, producing more accurate readings with fewer rays. This included moving the simplification away from the renderSweep function directly into the raySweep function.

All pygame related code has been refactored into it's own class and module for easier modularity and interfacing with different libraries (as per Brother Manley's instructions)

The foundations were laid for future expansion of the HUD, including shifting the viewport upwards and adding a bar.

Finally, We moved to the latest update of Pygame, to version 2, which added scaling functionality to make it less performance intensive to scale to a larger resolution for higher resolution displays.